+++ react_apps/src/components/loadingStudy.js
@@ -1,0 +1,6 @@
+const LoadingEl = () => {
+
+    return (<strong>Loding......</strong>);
+}
+
+export default LoadingEl;
