+++ react_apps/src/index.js
