+++ react_apps/src/routes/detailStudy.js
@@ -1,0 +1,8 @@
+function Detail() {
+
+    return (
+        <div>Detail</div>
+    )
+}
+
+export default Detail;
